begin refactoring jest tests to mocha+chai+sinon tests
