+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "node src/server.js",
-    "test": "NODE_ENV=test jest --no-cache --maxWorkers=2 --config=.jest.json"
+    "test": "./node_modules/.bin/mocha --reporter spec"
   },
   "author": "j@also.engineering",
   "repository": "assess-intl/elevator",
@@ -20,12 +20,14 @@
     "unirest": "^0.5.1"
   },
   "devDependencies": {
+    "chai": "^4.1.2",
     "coveralls": "^3.0.0",
     "eslint": "^4.18.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.9.0",
     "eslint-plugin-jest": "^21.12.2",
-    "jest": "^22.4.1",
+    "mocha": "^5.0.1",
+    "sinon": "^4.3.0",
     "supertest": "^3.0.0"
   }
 }
